requirements.txt
appdirs==1.4.3
certifi==2018.1.18
Django==1.10.6
docutils==0.13.1
packaging==16.8
pipenv==11.8.0
psutil==5.0.1
pyowm==2.8.0
Pygments==2.2.0
pyparsing==2.2.0
pyTelegramBotAPI==3.6.6
python-telegram-bot==7.0.1
requests==2.20.0
six==1.10.0
virtualenv==15.1.0
virtualenv-clone==0.3.0
wget==3.2
pydub==0.23.1
jsonschema==3.2.0
parser-json-python==1.0.0
psycopg2==2.8.3
psycopg2-binary==2.8.4
psycopg2-pool==1.1
Flask==1.1.1
Flask-Bcrypt==0.7.1
Flask-Caching==1.8.0
Flask-DebugToolbar==0.10.1
Flask-Login==0.4.1
Flask-Migrate==2.5.2
Flask-SQLAlchemy==2.4.1
Flask-Static-Digest==0.1.2
Flask-WTF==0.14.2
configparser==4.0.2
regex==2020.1.8
